requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
aenum=3.1.15=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.11.2=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
astroid=3.3.5=pypi_0
async-timeout=5.0.1=pypi_0
attrs=24.2.0=pypi_0
bidict=0.23.1=pypi_0
black=24.10.0=pypi_0
boto3=1.35.63=pypi_0
botocore=1.35.63=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.9.24=hca03da5_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.1.0=pypi_0
datasets=3.1.0=pypi_0
dill=0.3.8=pypi_0
evaluate=0.4.3=pypi_0
filelock=3.16.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
h11=0.14.0=pypi_0
huggingface-hub=0.26.2=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
isort=5.13.2=pypi_0
jinja2=3.1.4=pypi_0
jmespath=1.0.1=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
lightly=1.5.14=pypi_0
lightly-utils=0.0.2=pypi_0
lightning-utilities=0.11.8=pypi_0
markupsafe=3.0.2=pypi_0
mccabe=0.7.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h313beb8_0
networkx=3.2.1=pypi_0
numpy=1.26.4=pypi_0
omegaconf=2.3.0=pypi_0
opacus=1.4.0=pypi_0
openssl=3.0.15=h80987f9_0
opt-einsum=3.4.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pathspec=0.12.1=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py39hca03da5_0
platformdirs=4.3.6=pypi_0
plato-learn=1.0=pypi_0
propcache=0.2.0=pypi_0
pyarrow=18.0.0=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pylint=3.3.1=pypi_0
python=3.9.20=hb885b13_1
python-dateutil=2.9.0.post0=pypi_0
python-engineio=4.10.1=pypi_0
python-socketio=5.11.4=pypi_0
pytorch-lightning=2.2.5=pypi_0
pytorch-ranger=0.1.1=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h1a28f6b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
s3transfer=0.10.3=pypi_0
safetensors=0.4.5=pypi_0
scipy=1.13.1=pypi_0
setuptools=75.1.0=py39hca03da5_0
simple-websocket=1.1.0=pypi_0
six=1.16.0=pypi_0
sqlite=3.45.3=h80987f9_0
sympy=1.13.1=pypi_0
timm=1.0.11=pypi_0
tk=8.6.14=h6ba3021_0
tokenizers=0.20.3=pypi_0
tomli=2.1.0=pypi_0
tomlkit=0.13.2=pypi_0
torch=1.13.1=pypi_0
torch-optimizer=0.3.0=pypi_0
torchmetrics=1.5.2=pypi_0
torchstat=0.0.7=pypi_0
torchvision=0.14.1=pypi_0
tqdm=4.67.0=pypi_0
transformers=4.46.2=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=1.26.20=pypi_0
wheel=0.44.0=py39hca03da5_0
wsproto=1.2.0=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h80987f9_1
yarl=1.17.2=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h18a0788_1
zstd=1.5.5.1=pypi_0
